refactor: import 수정 및 미사용 메서드 제거
