+++ src/main/java/com/example/blog/model/mapper/ArticleMapper.java
@@ -1,31 +1,18 @@
 package com.example.blog.model.mapper;
 
-import com.example.blog.model.dto.ArticleDto;
-import com.example.blog.model.dto.ArticleEmbeddedResponseDto;
-import com.example.blog.model.dto.ArticleResponseDto;
-import com.example.blog.model.dto.MemberDto;
+import com.example.blog.model.dto.article.ArticleEmbeddedResponseDto;
+import com.example.blog.model.dto.article.ArticlePostRequestDto;
+import com.example.blog.model.dto.article.ArticleResponseDto;
+import com.example.blog.model.dto.member.MemberResponseDto;
 import com.example.blog.model.entity.Article;
 import com.example.blog.model.entity.Member;
 
 public class ArticleMapper {
-    public static Article toEntity(ArticleDto dto, MemberDto memberDto) {
-        if (dto == null) {
+    public static Article toEntityFromPostRequest(ArticlePostRequestDto articleDto, Member member) {
+        if (articleDto == null) {
             return null;
         }
-        Member member = MemberMapper.toEntity(memberDto);
-        return new Article(dto.getId(), member, dto.getTitle(), dto.getContents());
-    }
-
-    public static ArticleDto toDto(Article article) {
-        if (article == null) {
-            return null;
-        }
-        return new ArticleDto(
-                article.getId(),
-                article.getMember().getId(),
-                article.getTitle(),
-                article.getContents()
-        );
+        return new Article(articleDto.getId(), member, articleDto.getTitle(), articleDto.getContents());
     }
 
     public static ArticleResponseDto toResponseDto(Article article) {
@@ -33,9 +20,11 @@ public class ArticleMapper {
             return null;
         }
 
+        MemberResponseDto memberResponseDto = MemberMapper.toResponseDto(article.getMember());
+
         return new ArticleResponseDto(
                 article.getId(),
-                article.getMember().getId(),
+                memberResponseDto,
                 article.getTitle(),
                 article.getContents(),
                 article.getCreatedAt(),
